+++ lib/src/authentication/mod.rs
@@ -1,5 +1,4 @@
 pub mod file_based;
-pub mod radius;
 
 
 use std::borrow::Cow;